reorganized components, refactored APIs
