+++ src/app/api/addUser/route.ts
@@ -4,6 +4,15 @@ import { prisma } from "../../../../lib/prismaClient";
 export async function POST(request: Request) {        
     const body = await request.json();
 
+    try {
+        await addUser();
+        return new NextResponse('Added User', { status: 200 });
+
+    } catch(error) {
+        console.error("error");
+        return new NextResponse('Database Error', { status: 500 });
+    };
+
     async function addUser() {        
         await prisma.user.create({
             data: {
@@ -14,14 +23,5 @@ export async function POST(request: Request) {
                 isReady: false
             }
         });                  
-    };
-
-    try {
-        await addUser();
-        return new NextResponse('Added User', { status: 200 });
-
-    } catch(error) {
-        console.error("error");
-        return new NextResponse('Database Error', { status: 500 });
     };
 };
