+++ src/app/api/checkAllReady/route.ts
@@ -5,7 +5,7 @@ export async function POST(request: Request) {
     const body = await request.json();
 
     try {
-        const users = await findUsers(); 
+        const users = await findUsers(body.gameId); 
         const isReadyArray = users.map(user => user.isReady);          
         
         const isAllReady = isReadyArray.every((bool) => bool === true);
@@ -17,10 +17,10 @@ export async function POST(request: Request) {
         return new NextResponse('Database Error', { status: 500 });
     };
 
-    async function findUsers() {        
+    async function findUsers(gameId: string) {        
         const users = await prisma.user.findMany({
             where: {
-                gameId: body.gameId                
+                gameId: gameId                
             }
         });       
         return users;
