+++ src/app/api/updateUserIsReady/route.ts
@@ -6,13 +6,13 @@ export async function POST(request: Request) {
     const body = await request.json();
 
     try {
-        await updateUserIsReady(); 
+        await updateUserIsReady(body.readyStatus); 
         
         if (body.readyStatus === true) {
             await pusherServer.trigger(body.gameId, "checkAllReady", body.nextGamePeriod)
-            .catch((error: any) => {
-                console.log(error);
-            });
+                .catch((error: any) => {
+                    console.log(error);
+                });        
         };  
 
         return new NextResponse('Updated user ready status', { status: 200 });
@@ -22,13 +22,13 @@ export async function POST(request: Request) {
         return new NextResponse('DatabaseError', { status: 500 });
     };
     
-    async function updateUserIsReady() {        
+    async function updateUserIsReady(readyStatus: boolean) {        
         await prisma.user.update({
             where: {
                 id: body.userId 
             },
             data: {
-                isReady: body.readyStatus
+                isReady: readyStatus
             }
         });         
     };
