+++ src/app/components/Game/EndScreen.tsx
