+++ src/app/components/Game/Score.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useState, useEffect } from 'react';
-import { updateUserIsReady } from '../../../../../lib/utils';
+import { updateUserIsReady } from '../../../../lib/utils';
 
 
 interface Props {
